fix(context): remove all context menus before creating any
For context menu clicks that may start the worker we can't rely on
Chrome to update the menu item to the right state visually. When
clicked we also create the item with the right state to update it
manually but also queue the creation to ensure it happens after
the creation triggered by init and not before.